Cave generation and random spawnpoints

+++ src/main/kotlin/me/orange/game/world/generation/ChunkGenerator.kt
@@ -3,6 +3,6 @@ package me.orange.game.world.generation
 import me.orange.game.utils.Vec
 import me.orange.game.world.chunk.Chunk
 
-abstract class ChunkGenerator(val seed: Int) {
+abstract class ChunkGenerator(val seed: Long) {
     abstract fun generateChunk(vec: Vec): Chunk
 }
